// lib/shopify/types.ts
export type Menu = {
  title: string;
  path: string;
  children?: Menu[];
};